Make Photo processing a full-page studio
The modal wizard forced a linear path through settings that are actually
interdependent — tone and shading get judged against each other, so
stepping forward and back through a dialog was the wrong shape.

It is now a full-page workspace: controls on the left, a live preview of
whatever is being edited on the right. The steps become two tabs,
"Prepare Image" (tone) and "Style" (shading), switchable in either
direction. Until an image is loaded the workspace is replaced entirely by
a picker, which now also takes a drag-and-drop.

Strokes re-render on a debounce instead of on a button, and are only
built while the Style tab is open since that is the expensive step. Any
upstream change drops the previous render outright — otherwise editing
tone on the Prepare tab would leave stale strokes behind and "Create
project" would commit geometry from the settings before the edit.
